Move ociRef by SPDX (only consumer), move goRef/bomRef into mod*.go, refactor bomRef to match goRef change.
