+++ internal/sbom/mod_1.18.go
@@ -20,6 +20,7 @@ package sbom
 import (
 	"fmt"
 	"runtime/debug"
+	"strings"
 )
 
 type BuildInfo debug.BuildInfo
@@ -31,4 +32,30 @@ func ParseBuildInfo(data string) (*BuildInfo, error) {
 	}
 	bi := BuildInfo(*dbi)
 	return &bi, nil
+}
+
+func modulePackageName(mod *debug.Module) string {
+	return fmt.Sprintf("SPDXRef-Package-%s-%s",
+		strings.ReplaceAll(mod.Path, "/", "."),
+		mod.Version)
+}
+
+func bomRef(mod *debug.Module) string {
+	return fmt.Sprintf("pkg:golang/%s@%s?type=module", mod.Path, mod.Version)
+}
+
+func goRef(mod *debug.Module) string {
+	path := mod.Path
+	// Try to lowercase the first 2 path elements to comply with spec
+	// https://github.com/package-url/purl-spec/blob/master/PURL-TYPES.rst#golang
+	p := strings.Split(path, "/")
+	if len(p) > 2 {
+		path = strings.Join(
+			append(
+				[]string{strings.ToLower(p[0]), strings.ToLower(p[1])},
+				p[2:]...,
+			), "/",
+		)
+	}
+	return fmt.Sprintf("pkg:golang/%s@%s?type=module", path, mod.Version)
 }
